refactor(tools): rename *_pi tool builtins + registry to canonical (M6.4)
- Rename cubebox/tools/builtin/{calculator,datetime_tool,view_images,memory,
  load_skill}_pi.py and cubebox/tools/registry_pi.py to canonical names.
- Rename factory functions: create_memory_tools_pi → create_memory_tools,
  create_load_skill_tool_pi → create_load_skill_tool,
  list_builtin_tools_for_cubepi → list_builtin_tools.
- Update imports in run_manager, skills middleware, tools/__init__,
  tools/builtin/__init__, and all unit tests.
- Rename test files tests/unit/test_*_pi.py → tests/unit/test_*.py.
