+++ backend/cubebox/tools/registry.py
@@ -6,7 +6,7 @@ M2.4 adds MCP loading; M3 wires per-conversation dynamic tool sets.
 Note: view_images requires per-request DI (org_id, workspace_id, objectstore,
 capabilities) and is therefore NOT included in the no-DI list. Callers that
 need view_images should call make_view_images_tool(...) from
-cubebox.tools.builtin.view_images_pi and compose it into the tool list at
+cubebox.tools.builtin.view_images and compose it into the tool list at
 agent-construction time.
 """
 
@@ -14,11 +14,11 @@ from __future__ import annotations
 
 from cubepi.agent.types import AgentTool
 
-from cubebox.tools.builtin.calculator_pi import calculator_tool
-from cubebox.tools.builtin.datetime_tool_pi import datetime_tool
+from cubebox.tools.builtin.calculator import calculator_tool
+from cubebox.tools.builtin.datetime_tool import datetime_tool
 
 
-def list_builtin_tools_for_cubepi() -> list[AgentTool]:  # type: ignore[type-arg]
+def list_builtin_tools() -> list[AgentTool]:  # type: ignore[type-arg]
     """Return cubepi-shaped builtin tools that require no DI.
 
     M2.2/M2.3 will extend with memory + load_skill (factory functions
@@ -26,7 +26,7 @@ def list_builtin_tools_for_cubepi() -> list[AgentTool]:  # type: ignore[type-arg
 
     view_images is intentionally excluded; it needs org_id, workspace_id,
     objectstore, and LLMCapabilities. Use make_view_images_tool(...) from
-    cubebox.tools.builtin.view_images_pi when constructing an agent that
+    cubebox.tools.builtin.view_images when constructing an agent that
     handles image attachments.
     """
     return [calculator_tool, datetime_tool]
